Add Dword and DwordByReference types.
With the Dword type that is always 32 bits on Windows and Mac but long on Linux, we can unify the function and struct declarations that we give to JNA.

The downside is that Dword must have a size that is statically computed. If you were to compile your own version of pcsclite, you would need to modify the Dword static size. I do not think that is likely to be needed though.